+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [0.11.5] (Unreleased)
 - Fix: non-total typemappers now work in `no_box` fields (#1198)
+- Memoize serializedSize of empty messages (previously serialized size was recomputed for
+  empty message)
 
 ## [0.11.4]
 - Add options for enum annotations (#1155)